feat: rename accounts to bank-accounts
