--- src/core/accounts/BalanceCard/index.ts
@@ -1,1 +1,0 @@
-export * from './BalanceCard'
