--- src/core/accounts/index.ts
@@ -1,1 +1,0 @@
-export * from './BalanceCard'
