+++ src/core/bank-accounts/BankAccountBalanceCard/index.ts
@@ -1,0 +1,1 @@
+export * from './BankAccountBalanceCard'
