+++ src/core/bank-accounts/BankAccountsList/index.ts
@@ -1,0 +1,1 @@
+export * from './BankAccountsList'
