+++ src/core/bank-accounts/index.ts
@@ -1,0 +1,2 @@
+export * from './BankAccountBalanceCard'
+export * from './BankAccountsList'
