+++ src/lib/router/index.ts
@@ -1,6 +1,6 @@
 export * from './Router'
 
 export const SIGN_IN = '/'
-export const ACCOUNTS = '/accounts'
+export const ACCOUNTS = '/bank_accounts'
 export const DASHBOARD = '/dashboard'
 export const CREATE_ACCOUNT = '/create_account'
