--- src/pages/Accounts/index.ts
@@ -1,1 +1,0 @@
-export * from './Accounts'
