+++ src/pages/BankAccounts/index.ts
@@ -1,0 +1,1 @@
+export * from './BankAccounts'
