+++ src/pages/index.ts
@@ -1,4 +1,4 @@
 export * from './SignIn'
 export * from './Dashboard'
 export * from './CreateAccount'
-export * from './Accounts'
+export * from './BankAccounts'
